refactor: input과 로또 생성도 별도의 파일로 분리
